[alpha] BossStateWaveTest : start the wave from the state.
