+++ code/Alpha/Components/Boss/States/BossStateWaveTest.h
@@ -6,13 +6,29 @@
 
 #include "Alpha/StateMachine/IState.h"
 
+namespace Systems
+{
+	class MeshAsset;
+	class MaterialInstanceAsset;
+}
+
+class BossComponent;
+class WaveTest;
+
 class BossStateWaveTest : public IState
 {
 public:
 	BossStateWaveTest(StateMachine* pStateMachine);
 	~BossStateWaveTest();
 
+	void Init(Systems::MeshAsset* pMesh, Systems::MaterialInstanceAsset* pMaterial, BossComponent* pComponent);
+
 	void OnEnter() override;
 	void OnUpdate() override;
 	void OnExit() override;
+
+private:
+	WaveTest* m_pWave;
+	BossComponent* m_pBossComponent;
+	uint32_t m_waveIndex;
 };
